Refactorize deployment code
- Split the massive deploy.rb model into smaller files for readability sake.
- Change the way ASYD has to read def files to keep track of the line it's
being executed, and jump to a concrete line at will.
